add app teacher move teachers models to teachers app
