+++ backend/apps/students/models.py
@@ -7,14 +7,4 @@ class Student(models.Model):
     birth_date = models.IntegerField()
     gender = models.CharField(max_length=10, choices=(('male', 'male'), ('female', 'female')))
 
-class Teacher(models.Model) :
-    teacher_id = models.IntegerField(primary_key=True)
-    passwd = models.CharField(max_length=100)
-    age = models.IntegerField()
-    name = models.CharField(max_length=100)
-    position = models.CharField(max_length=100)
-    sex = models.CharField(max_length=100)
-
-    def __str__(self):
-        return str(self.teacher_id)
